apps/polycode/pyproject.toml
[project]
name = "coding_farm"
version = "0.1.0"
description = "Antfarm workflows implemented with CrewAI - Multi-agent software development automation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.4.0",
    "click>=8.0.0",
    "crewai[litellm]>=1.10.0b1",
    "crewai-tools>=0.76.0",
    "fastapi>=0.115.0",
    "flower>=2.0.1",
    "gitpython>=3.1.0",
    "httpx>=0.27.0",
    "ollama>=0.6.1",
    "pathspec>=0.12.0",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.0.0",
    "pygithub>=2.8.1",
    "pyjwt[crypto]>=2.9.0",
    "python-dotenv>=1.0.0",
    "redis>=5.2.0",
    "sqlalchemy>=2.0.0",
    "sqlmodel>=0.0.22",
    "ty>=0.0.21",
    "uvicorn[standard]>=0.32.0",
    "python-socketio>=5.16.1",
    "langchain-core>=1.2.19",
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "tree-sitter-go>=0.23.0",
    "pygls>=1.3.0",
    "typer>=0.23.1",
    "rich>=14.3.3",
    "tqdm>=4.67.3",
    "alive-progress==3.3.0",
]

[project.scripts]
polycode = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/bug_fix",
    "src/github_app",
    "src/github_issues",
    "src/project_manager",
    "src/persistence",
    "src/celery_tasks",
    "src/cli",
]

[dependency-groups]
dev = ["pytest>=9.0.2", "ruff>=0.15.4"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pycodestyle]
count = false
ignore = ["E226", "E302", "E71"]
max-line-length = 160
statistics = true

[tool.pytest.ini_options]
# pythonpath = ["src"]
pythonpath = ["./src"]
addopts = [
    "-v",                      # verbose output
    "--strict-markers",        # strict markers
    "--tb=short",              # shorter tracebacks
    "--maxfail=1",             # stop after first failure
    "-ra",                     # show summary of all tests
    "--import-mode=importlib",
]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
